requirements.txt
pypng

pyqrcode